Rjesene vjezbe 3

+++ Vjezbe_3/ElementaryParticle.h
@@ -1,16 +1,19 @@
 #include <iostream>
 #include <cmath>
+#include <fstream>
+#include <cstdlib>
+#include <ctime>
 
 using namespace std;
 
 class ElementaryParticle{
 	public:
 		string ime; 
-		int masa;
+		double masa;
 		bool jeBozon;
 		double px, py, pz, E;
 		
-		ElementaryParticle(string n, int m, bool b, double p_x, double p_y, double p_z);
+		ElementaryParticle(string n, double m, bool b, double p_x, double p_y, double p_z);
 		void printInfo();
 		void Momentum(double kolx, double koly, double kolz);
 		ElementaryParticle();
